Initial changes for SQLAlchemy

+++ test_datatable.py
@@ -15,7 +15,10 @@ from c12n_pipe.io.data_catalog import AbstractDataTable, DataCatalog
 from c12n_pipe.io.node import Pipeline, StoreNode, PythonNode, LabelStudioNode
 from sqlalchemy.sql.sqltypes import JSON
 
-DBCONNSTR = f'postgresql://postgres:password@{os.getenv("POSTGRES_HOST", "localhost")}:{os.getenv("POSTGRES_PORT", 5432)}/postgres'
+# DBCONNSTR = f'postgresql://postgres:password@{os.getenv("POSTGRES_HOST", "localhost")}:{os.getenv("POSTGRES_PORT", 5432)}/postgres'
+DBCONNSTR = 'sqlite:///:memory:'
+DB_TEST_SCHEMA = None
+DB_CREATE_SCHEMA = False
 
 TEST_SCHEMA = [
     Column('a', Numeric),
@@ -43,18 +46,22 @@ def yield_df(data):
 
 @pytest.fixture
 def test_schema():
-    eng = create_engine(DBCONNSTR)
+    if DB_CREATE_SCHEMA:
+        eng = create_engine(DBCONNSTR)
 
-    try:
+        try:
+            eng.execute('DROP SCHEMA test CASCADE')
+        except:
+            pass
+
+        eng.execute('CREATE SCHEMA test')
+
+        yield 'ok'
+
         eng.execute('DROP SCHEMA test CASCADE')
-    except:
-        pass
-
-    eng.execute('CREATE SCHEMA test')
-
-    yield 'ok'
-
-    eng.execute('DROP SCHEMA test CASCADE')
+    
+    else:
+        yield 'ok'
 
 
 def check_df_equal(a, b):
@@ -81,7 +88,7 @@ def assert_idx_equal(a, b):
 
 @pytest.mark.usefixtures('test_schema')
 def test_cloudpickle():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl = ds.get_table('test', TEST_SCHEMA)
 
@@ -90,7 +97,7 @@ def test_cloudpickle():
 
 @pytest.mark.usefixtures('test_schema')
 def test_simple():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl = ds.get_table('test', TEST_SCHEMA)
 
@@ -99,7 +106,7 @@ def test_simple():
 
 @pytest.mark.usefixtures('test_schema')
 def test_store_less_values():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl = ds.get_table('test', TEST_SCHEMA)
 
@@ -112,7 +119,7 @@ def test_store_less_values():
 
 @pytest.mark.usefixtures('test_schema')
 def test_get_process_ids():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
     tbl2 = ds.get_table('tbl2', TEST_SCHEMA)
@@ -135,7 +142,7 @@ def test_get_process_ids():
 
 @pytest.mark.usefixtures('test_schema')
 def test_gen_process():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl1_gen = ds.get_table('tbl1_gen', TEST_SCHEMA)
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
@@ -181,7 +188,7 @@ def test_gen_process():
 
 @pytest.mark.usefixtures('test_schema')
 def test_inc_process_modify_values() -> None:
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
     tbl2 = ds.get_table('tbl2', TEST_SCHEMA)
@@ -205,7 +212,7 @@ def test_inc_process_modify_values() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_inc_process_delete_values_from_input() -> None:
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
     tbl2 = ds.get_table('tbl2', TEST_SCHEMA)
@@ -229,7 +236,7 @@ def test_inc_process_delete_values_from_input() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_inc_process_delete_values_from_proc() -> None:
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
     tbl2 = ds.get_table('tbl2', TEST_SCHEMA)
@@ -248,7 +255,7 @@ def test_inc_process_delete_values_from_proc() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_inc_process_proc_no_change() -> None:
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
     tbl2 = ds.get_table('tbl2', TEST_SCHEMA)
@@ -279,7 +286,7 @@ def test_inc_process_proc_no_change() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_gen_process_many():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl_gen = ds.get_table('tbl_gen', TEST_SCHEMA)
     tbl1_gen = ds.get_table('tbl1_gen', TEST_SCHEMA)
@@ -318,7 +325,7 @@ def test_gen_process_many():
 
 @pytest.mark.usefixtures('test_schema')
 def test_inc_process_many_modify_values() -> None:
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     tbl = ds.get_table('tbl', TEST_SCHEMA)
     tbl1 = ds.get_table('tbl1', TEST_SCHEMA)
@@ -372,7 +379,7 @@ def test_inc_process_many_modify_values() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_inc_process_many_several_outputs():
-    ds = DataStore(DBCONNSTR, schema='test')
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
 
     BAD_IDXS = ['id_0', 'id_1', 'id_5', 'id_8']
     GOOD_IDXS = ['id_2', 'id_3', 'id_4', 'id_6', 'id_7', 'id_9']
@@ -402,12 +409,12 @@ def test_inc_process_many_several_outputs():
 
 @pytest.mark.usefixtures('test_schema')
 def test_data_catalog() -> None:
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
     data_catalog = DataCatalog(
+        ds=ds,
         catalog={
             'test_df': AbstractDataTable([Column('a', Numeric)])
         },
-        connstr=DBCONNSTR,
-        schema='test_schema'
     )
 
     tbl = data_catalog.get_data_table('test_df')
@@ -417,12 +424,12 @@ def test_data_catalog() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_store_node() -> None:
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
     data_catalog = DataCatalog(
+        ds=ds,
         catalog={
             'test_df': AbstractDataTable([Column('a', Numeric)])
         },
-        connstr=DBCONNSTR,
-        schema='test_schema'
     )
 
     def proc_func():
@@ -440,15 +447,15 @@ def test_store_node() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_python_node() -> None:
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
     data_catalog = DataCatalog(
+        ds=ds,
         catalog={
             'test_df': AbstractDataTable([Column('a', Numeric)]),
             'test_df_inc1': AbstractDataTable([Column('a', Numeric)]),
             'test_df_inc2': AbstractDataTable([Column('a', Numeric)]),
             'test_df_inc3': AbstractDataTable([Column('a', Numeric)])
         },
-        connstr=DBCONNSTR,
-        schema='test_schema'
     )
     data_catalog.get_data_table('test_df').store(TEST_DF)
 
@@ -490,13 +497,13 @@ def temp_dir():
 
 @pytest.mark.usefixtures('test_schema', 'temp_dir')
 def test_label_studio_node() -> None:
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
     data_catalog = DataCatalog(
+        ds=ds,
         catalog={
             'data_input': AbstractDataTable([Column('data', JSON)]),
             'data_annotation': AbstractDataTable([Column('data', JSON)])
         },
-        connstr=DBCONNSTR,
-        schema='test_schema'
     )
 
     label_studio_node = LabelStudioNode(
@@ -512,15 +519,15 @@ def test_label_studio_node() -> None:
 
 @pytest.mark.usefixtures('test_schema')
 def test_pipeline() -> None:
+    ds = DataStore(DBCONNSTR, schema=DB_TEST_SCHEMA)
     data_catalog = DataCatalog(
+        ds=ds,
         catalog={
             'test_df': AbstractDataTable([Column('a', Numeric)]),
             'test_df_inc1': AbstractDataTable([Column('a', Numeric)]),
             'test_df_inc2': AbstractDataTable([Column('a', Numeric)]),
             'test_df_inc3': AbstractDataTable([Column('a', Numeric)])
         },
-        connstr=DBCONNSTR,
-        schema='test_schema'
     )
 
     def store_func():
